Detect dbt project dir
