+++ internal/pkg/dbt/dbt.go
@@ -1,0 +1,3 @@
+package dbt
+
+const ProjectFile = "dbt_project.yml"
